re-baselining the scripts: tidying up makefile, using env vars & using shared common.py script
