+++ config.yaml
@@ -1,6 +1,7 @@
 ---
 input_tables:
     - bkp_change_history
+    - dim_business_unit
 other_params:
     gx_data_src_name: gx_datasource_snowflake
-    row_count_limit: 3
+    row_count_limit: 2
