--- gx/checkpoints/my_checkpoint.yml
@@ -1,31 +1,0 @@
----
-name: my_checkpoint
-config_version: 1.0
-template_name:
-module_name: great_expectations.checkpoint
-class_name: Checkpoint
-run_name_template:
-expectation_suite_name:
-batch_request: {}
-action_list:
-    - name: store_validation_result
-      action:
-          class_name: StoreValidationResultAction
-    - name: store_evaluation_params
-      action:
-          class_name: StoreEvaluationParametersAction
-    - name: update_data_docs
-      action:
-          class_name: UpdateDataDocsAction
-evaluation_parameters: {}
-runtime_configuration: {}
-validations:
-    - batch_request:
-          datasource_name: gx_datasource_snowflake
-          data_asset_name: bkp_change_history
-          options: {}
-          batch_slice:
-      expectation_suite_name: 20231008_expectations_bkp_change_history
-profilers: []
-ge_cloud_id:
-expectation_suite_ge_cloud_id:
