--- gx/uncommitted/validations/.ge_store_backend_id
@@ -1,1 +1,0 @@
-store_backend_id = 7b67c716-5946-4703-a9eb-8bcf2048e33a
